Implement key derivation for channel end
