+++ cairo-contracts/packages/core/src/tests/keys.cairo
@@ -1,4 +1,4 @@
-use starknet_ibc_core::host::{commitment_key, next_sequence_ack_key};
+use starknet_ibc_core::host::{channel_end_key, commitment_key, next_sequence_ack_key};
 use starknet_ibc_testkit::dummies::{CHANNEL_ID, PORT_ID, SEQUENCE};
 
 #[test]
@@ -19,3 +19,13 @@ fn test_commitment_key() {
     let expected_key: felt252 = 0x21d9da1890ea380cfe5af2c3e84da497be6d2c220159f4e81dd7949cf5512f1;
     assert_eq!(expected_key, key);
 }
+
+#[test]
+fn test_channel_end_key() {
+    let channel_id = CHANNEL_ID(0);
+    let port_id = PORT_ID();
+    let key = channel_end_key(@port_id, @channel_id);
+    let expected_key: felt252 = 0x4763930c1c148bddb94c2603126c8d849bbc6921a0fc46734f21ebd2016579e;
+    assert_eq!(expected_key, key);
+}
+
